+++ src/editors/query/components/ParserType.tsx
@@ -47,7 +47,6 @@ export const ParseTypeEditor = (props: { query: InfinityQuery; onChange: (value:
           options={[
             { value: 'simple', label: 'Default' },
             { value: 'backend', label: 'Backend' },
-            // { value: 'sqlite', label: 'SQLite' },
             { value: 'uql', label: 'UQL' },
             { value: 'groq', label: 'GROQ' },
           ]}
@@ -58,9 +57,6 @@ export const ParseTypeEditor = (props: { query: InfinityQuery; onChange: (value:
             } else if (query.parser === 'groq') {
               let groq = query.groq || '*';
               onChange({ ...query, parser: e.value, groq });
-            } else if (query.parser === 'sqlite') {
-              let sqlite_query = query.sqlite_query || 'SELECT * FROM input';
-              onChange({ ...query, parser: e.value, sqlite_query });
             } else {
               onChange({ ...query, parser: e.value });
             }